Fix out-of-bounds access in Puzzle::eval
There was some confusion about the sizes of some arrays. This started
with `indexToLetter` and `leading` in the constructor. They're don't
necessarily have as many entries as we have digits. We start with the
former empty (but reserve storage), and initialize the latter once we
know the proper size.

The `assignment` has than as many entries as we have letters, not as we
have digits. (The number of letters is used as domain size of the map.)